requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.0.1
django-redis==4.10.0
djangorestframework==3.10.3
idna==2.8
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.8
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
